// src/interface.ts
export interface CrewDataInterface {
    profession: string;
    name: string;
    details: string;
    image: string;
}

export interface DestinationDataInterface {
    planet: string;
    details: string;
    distance: string;
    time: string;
    image: string;
}

export interface TechnologyDataInterface {
    num: string;
    header: string;
    details: string;
    image: string;
}